requirements.txt
torch==2.5.1
torchvision==0.20.1
deepspeed
toml
transformers
diffusers>=0.32.1
datasets
pillow
sentencepiece
protobuf
peft
torch-optimi
tensorboard
tqdm
safetensors
bitsandbytes
imageio[ffmpeg]
av
einops
accelerate
loguru
flash-attn
omegaconf
iopath
termcolor
hydra-core
gradio==5.22
wandb
sageattention==1.0.6
easydict
ftfy
pytorch-optimizer
optimum-quanto